% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.externaltrade1993.R
\docType{data}
\name{gh.externaltrade1993}
\alias{gh.externaltrade1993}
\alias{gh.externaltrade1993-package}
\title{Data of external trade statistics of Ghana in 1993}
\format{
A tibble with 6268 rows and 6 variables:
\describe{
\item{partner}{chr denotes trading partner countries or origin of goods}
\item{description}{dbl denotes the item being imported}
\item{cif_value_cedi}{dbl denoting Cost-Insurance-Freight value of goods to importer}
\item{net_weight_kg}{dbl denoting net weight of goods exclusive of packings and wrappings}
\item{net_weight_cedi_kg}{dbl denoting relative value of goods in cedi per weight}
\item{share_percent}{dbl denoting the ratio of CIF value to grand total CIF  value of all imports in percent}
}
}
\usage{
gh.externaltrade1993
}
\description{
The detailed statistical information on the external trade statistics of Ghana from January to June 1993. Published by Ghana Statistical Service, the variables are given below:
}
\author{
Paul Agbodza
}
\keyword{datasets}
